Switch drive in MenuViewController
