+++ kDrive/UI/Controller/Home/SelectSwitchDriveDelegate.swift
@@ -1,0 +1,45 @@
+/*
+ Infomaniak kDrive - iOS App
+ Copyright (C) 2021 Infomaniak Network SA
+
+ This program is free software: you can redistribute it and/or modify
+ it under the terms of the GNU General Public License as published by
+ the Free Software Foundation, either version 3 of the License, or
+ (at your option) any later version.
+
+ This program is distributed in the hope that it will be useful,
+ but WITHOUT ANY WARRANTY; without even the implied warranty of
+ MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ GNU General Public License for more details.
+
+ You should have received a copy of the GNU General Public License
+ along with this program.  If not, see <http://www.gnu.org/licenses/>.
+ */
+
+import kDriveCore
+import UIKit
+
+protocol SelectSwitchDriveDelegate: SelectDelegate, UIViewController {}
+extension SelectSwitchDriveDelegate {
+    func didSelect(option: Selectable) {
+        if let drive = option as? Drive {
+            if drive.maintenance {
+                let driveFloatingPanelController = DriveMaintenanceFloatingPanelViewController.instantiatePanel()
+                let floatingPanelViewController = driveFloatingPanelController.contentViewController as? DriveMaintenanceFloatingPanelViewController
+                floatingPanelViewController?.setTitleLabel(with: drive.name)
+                present(driveFloatingPanelController, animated: true)
+            } else {
+                AccountManager.instance.setCurrentDriveForCurrentAccount(drive: drive)
+                AccountManager.instance.saveAccounts()
+                // Download root file
+                guard let currentDriveFileManager = AccountManager.instance.currentDriveFileManager else {
+                    return
+                }
+
+                currentDriveFileManager.getFile(id: DriveFileManager.constants.rootID) { [weak self] _, _, _ in
+                    (self?.tabBarController as? SwitchDriveDelegate)?.didSwitchDriveFileManager(newDriveFileManager: currentDriveFileManager)
+                }
+            }
+        }
+    }
+}
